+++ tools/tbench/w11a_ibtst/test_ibtst_data.tcl
@@ -1,10 +1,11 @@
-# $Id: test_ibtst_data.tcl 1112 2019-02-17 11:10:04Z mueller $
+# $Id: test_ibtst_data.tcl 1120 2019-03-09 18:19:31Z mueller $
 #
 # Copyright 2019- by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 # License disclaimer see License.txt in $RETROBASE directory
 #
 # Revision History:
 # Date         Rev Version  Comment
+# 2019-03-05  1118   1.0.1  use -wal
 # 2019-02-16  1112   1.0    Initial version
 #
 # Test data register response
@@ -20,7 +21,7 @@
 rlc log "test_ibtst_data: test data register response -------------------------"
 
 if {[$cpu get hasibtst] == 0} {
-  rlc log "  test_ibtst_data-W: no ibtest unit found, test aborted"
+  rlc log "  test_ibtst_data-W: no ibtst unit found, test aborted"
   return
 }
 package require ibd_ibtst
@@ -29,7 +30,7 @@ rlc log "    A1: data loc/rem access ----------------------------"
 # test loc/rem write: n/n; y/n; n/y; y/y
 # test loc/rem read:  n/n; y/n; n/y; y/y
 $cpu cp \
-  -wa [cpu0 imap  it.data] \
+  -wal [$cpu imap  it.data] \
   -wreg it.cntl [regbld ibd_ibtst::CNTL ] \
   -wm           0xdead -estaterr \
   -wibr it.data 0xdead -estaterr \
@@ -55,9 +56,15 @@ $cpu cp \
   -rm           -edata 0xdead           \
   -ribr it.data -edata 0xdead
 
-rlc log "    A2: data loc nak and timeout -----------------------"
-# test datto; should look like nak 
+rlc log "    A2: data loc nak and bsy,bsy+ack,timeout -----------"
+# should all give rbus
 $cpu cp \
+  -wreg it.cntl [regbld ibd_ibtst::CNTL bsyw bsyr ] \
+  -wm 0xdead -estaterr \
+  -rm        -estaterr \
+  -wreg it.cntl [regbld ibd_ibtst::CNTL bsyw bsyr datab ] \
+  -wm 0xdead -estaterr \
+  -rm        -estaterr \
   -wreg it.cntl [regbld ibd_ibtst::CNTL datto ] \
   -wm 0xdead -estaterr \
   -rm        -estaterr
@@ -92,3 +99,7 @@ $cpu cp \
   -wibr it.data 0xdead \
   -breset \
   -ribr it.data -edata 0x0000
+
+# harvest breset/creset triggered attn's
+rlc exec -attn -edata 0
+rlc wtlam 0.